Separate hangman test
